feat!: rig-rmcp — move MCP tool support out of rig-agent into its own crate
Bevy-prep follow-up. rig-agent no longer has an rmcp feature or an rmcp
dependency; MCP tool support lives in the new rig-rmcp crate, mirroring
rig-reqwest: the SDK-bound integration depends on rig-agent/rig-core,
implements rig's traits for its own types, supplies ergonomics through
extension traits, and is re-exported by the rig facade behind the
existing rmcp feature. rig-agent's only optional tokio consumer is now
discord-bot; its default-feature dependency tree has no tokio and no
rmcp.

rig-agent
- The registry support for externally managed tools is public and
  ungated — nothing about "a remote tool list that can refresh or
  disconnect" is MCP-specific: ErasedTool (and its is_live hook),
  ToolSet::add_erased, ToolServer::erased_tool, AgentBuilder::erased_tool
  (both states), ToolServerHandle::{add_managed_erased_tools,
  reconcile_managed_erased_tools}, the opaque ManagedToolToken, and
  Agent::tool_server_handle(). retire_disconnected_tools now always runs.
- The rmcp_tool* methods on AgentBuilder and ToolServer, the tool/rmcp
  module, the wasm compile_error!, the rmcp feature and dependency are
  removed; tokio stays optional (discord-bot, test-utils).

rig-rmcp (new), two layers
- Core (default-features = false; rig-core only): McpTool
  (from_mcp_server, with_timeout, execute_mcp), tools_from_server,
  McpClientError, DEFAULT_MCP_TOOL_TIMEOUT, the result→output mapping,
  and From<McpTool> for rig_core::tool::PortableDynamicTool — MCP tools
  usable from any runtime that consumes portable tools, minus the two
  context-dependent conveniences (_meta passthrough, preserved raw
  result). tests/portable.rs drives one against an in-process rmcp
  server without rig-agent; cargo tree --no-default-features has no
  rig-agent.
- agent feature (default on): impl ErasedTool for McpTool (Meta
  passthrough from the ToolContext, preserved CallToolResult),
  McpClientHandler driving the now-public managed-registration API,
  RmcpAgentBuilderExt / RmcpToolServerExt (the former rmcp_tool*
  methods, verbatim) and a prelude. The in-process rmcp test servers
  and the builder timeout test move here.
- Native-only: the crate root raises the same compile_error! on wasm and
  gates its body, so a wasm check emits exactly one error.

Facade / workspace
- rig: rmcp = ["agent", "dep:rig-rmcp"]; rig::tool::rmcp re-exports
  rig_rmcp::*, rig::prelude carries rig_rmcp::prelude, so prelude users
  see no change. CI's "rmcp rejected on wasm" job targets rig-rmcp;
  rig-rmcp joins the nextest package list; provider_layout classifies it
  as an exempt package. MIGRATING.md + READMEs updated.

Verification: cargo check/clippy --workspace --all-features --all-targets
clean; rig-rmcp 30+1 tests (both feature sets), rig-agent 568, facade
987 tests pass; doctests and rustdoc -D warnings clean; CI's
wasm check of rig-rmcp fails with exactly one error (the compile_error!).
