+++ Cargo.toml
@@ -107,6 +107,7 @@ edition = "2024"
 rig = { path = ".", version = "0.42.0" }
 rig-agent = { path = "crates/rig-agent", version = "0.42.0" }
 rig-reqwest = { path = "crates/rig-reqwest", version = "0.42.0" }
+rig-rmcp = { path = "crates/rig-rmcp", version = "0.42.0" }
 rig-core = { path = "crates/rig-core", version = "0.42.0" }
 anyhow = "1"
 arrow-array = "58"
@@ -232,6 +233,7 @@ workspace = true
 rig-core = { path = "crates/rig-core", version = "0.42.0", default-features = false }
 rig-reqwest = { path = "crates/rig-reqwest", version = "0.42.0", optional = true, default-features = false }
 rig-agent = { path = "crates/rig-agent", version = "0.42.0", optional = true, default-features = false }
+rig-rmcp = { path = "crates/rig-rmcp", version = "0.42.0", optional = true }
 rig-bedrock = { path = "crates/rig-bedrock", version = "0.42.0", optional = true, default-features = false }
 rig-candle = { path = "crates/rig-candle", version = "0.42.0", optional = true, default-features = false }
 rig-fastembed = { path = "crates/rig-fastembed", version = "0.42.0", optional = true, default-features = false }
@@ -363,7 +365,7 @@ discord-bot = ["agent", "rig-agent/discord-bot"]
 pdf = ["rig-core/pdf"]
 epub = ["rig-core/epub"]
 rayon = ["rig-core/rayon"]
-rmcp = ["agent", "rig-agent/rmcp"]
+rmcp = ["agent", "dep:rig-rmcp"]
 socks = ["reqwest", "rig-reqwest/socks"]
 reqwest = ["dep:rig-reqwest"]
 websocket = ["reqwest", "rig-reqwest/websocket"]
