+++ README.md
@@ -72,7 +72,7 @@ More information about this crate can be found in the [official](https://rig.rs/
 - Integrate LLMs in your app with minimal boilerplate
 - Browser-WASM (`wasm32-unknown-unknown`) support for the portable core and
   classic runtime — see [target support](crates/rig-agent/README.md#target-support)
-  for the full matrix (WASI is not supported; `rmcp` is native-only)
+  for the full matrix (WASI is not supported; `rig-rmcp`/MCP is native-only)
 
 ## Runtime choices
