+++ crates/rig-agent/src/tool/mod.rs
@@ -125,27 +125,6 @@ use crate::completion::{self, ToolDefinition};
 
 pub(crate) mod extensions;
 
-// MCP is native-only. rmcp's `ClientHandler` is declared
-// `Sized + Send + Sync + 'static` unconditionally — its `local` feature relaxes
-// the future bounds (`MaybeSendFuture`) but not the handler itself — and this
-// crate's handler owns the tool registry, whose `Arc<dyn ErasedTool>` is
-// deliberately neither `Send` nor `Sync` on wasm because `rig-core`'s
-// `WasmCompatSend`/`WasmCompatSync` are no-op markers there. The two
-// maybe-`Send` abstractions cannot be reconciled from this side.
-//
-// Raise that as one sentence instead of a page of `dyn ErasedTool` trait errors.
-// Upstream fix would be making rmcp's handler bound conditional on `local`, as
-// its future bound already is.
-#[cfg(all(feature = "rmcp", target_family = "wasm"))]
-compile_error!(
-    "the `rmcp` feature is native-only: rmcp's `ClientHandler` requires \
-     `Send + Sync` unconditionally (its `local` feature relaxes only futures), \
-     which rig's wasm tool registry cannot satisfy. Disable `rmcp` for wasm targets."
-);
-
-#[cfg(all(feature = "rmcp", not(target_family = "wasm")))]
-#[cfg_attr(docsrs, doc(cfg(feature = "rmcp")))]
-pub mod rmcp;
 pub mod server;
 
 pub use extensions::{MissingToolContext, ToolContext};
@@ -304,7 +283,7 @@ where
 }
 
 /// Crate-private, object-safe dispatch boundary.
-pub(crate) trait ErasedTool: WasmCompatSend + WasmCompatSync {
+pub trait ErasedTool: WasmCompatSend + WasmCompatSync {
     fn name(&self) -> String;
     fn description(&self) -> String;
     fn parameters(&self) -> serde_json::Value;
@@ -312,7 +291,6 @@ pub(crate) trait ErasedTool: WasmCompatSend + WasmCompatSync {
     ///
     /// In-process tools are always live. Remote adapters override this so the
     /// registry can retire disconnected owners without probing by execution.
-    #[cfg(all(feature = "rmcp", not(target_family = "wasm")))]
     fn is_live(&self) -> bool {
         true
     }
@@ -538,7 +516,6 @@ impl RegisteredTool {
         definition_with_name(name, self.erased())
     }
 
-    #[cfg(all(feature = "rmcp", not(target_family = "wasm")))]
     pub(crate) fn is_live(&self) -> bool {
         self.erased().is_live()
     }
@@ -671,8 +648,9 @@ impl ToolSet {
         self.insert(RegisteredTool::Embedding(Arc::new(tool)))
     }
 
-    #[cfg(all(feature = "rmcp", not(target_family = "wasm")))]
-    pub(crate) fn add_erased(&mut self, tool: Arc<dyn ErasedTool>) -> String {
+    /// Register a pre-erased tool. The extension point for adapters that
+    /// implement [`ErasedTool`] directly (remote tool protocols such as MCP).
+    pub fn add_erased(&mut self, tool: Arc<dyn ErasedTool>) -> String {
         self.insert(RegisteredTool::Static(tool))
     }
