+++ crates/rig-rmcp/Cargo.toml
@@ -1,0 +1,46 @@
+[package]
+name = "rig-rmcp"
+version.workspace = true
+edition = { workspace = true }
+license = "MIT"
+readme = "README.md"
+description = "MCP (Model Context Protocol) tool support for Rig via the rmcp SDK: MCP tools as portable rig-core tools, and (with the `agent` feature) live-refreshing registration into rig-agent's tool server."
+repository = "https://github.com/0xPlaygrounds/rig"
+
+[package.metadata.docs.rs]
+all-features = true
+rustdoc-args = ["--cfg", "docsrs"]
+
+[lints]
+workspace = true
+
+[dependencies]
+rig-core = { path = "../rig-core", version = "0.42.0", default-features = false }
+rig-agent = { path = "../rig-agent", version = "0.42.0", default-features = false, optional = true }
+# MCP is native-only: rmcp's `ClientHandler` requires `Send + Sync`
+# unconditionally, which rig's wasm tool registry cannot satisfy. The crate
+# root raises a targeted `compile_error!` on wasm; scoping the SDK out of the
+# wasm graph keeps that the only error.
+futures = { workspace = true }
+serde_json = { workspace = true }
+thiserror = { workspace = true }
+tracing = { workspace = true }
+
+[target.'cfg(not(target_family = "wasm"))'.dependencies]
+rmcp = { workspace = true, features = ["client"] }
+tokio = { workspace = true, features = ["rt", "sync", "time"] }
+
+[dev-dependencies]
+rig-core = { path = "../rig-core", version = "0.42.0", default-features = false, features = ["test-utils"] }
+rig-agent = { path = "../rig-agent", version = "0.42.0", default-features = false, features = ["test-utils"] }
+rmcp = { workspace = true, features = ["client", "server"] }
+tokio = { workspace = true, features = ["full"] }
+tokio-test = { workspace = true }
+serde_json = { workspace = true }
+
+[features]
+default = ["agent"]
+# Contextual MCP tools (`ErasedTool` with `_meta` passthrough and preserved
+# raw results), the live-refreshing `McpClientHandler`, and the
+# `AgentBuilder` / `ToolServer` extension traits.
+agent = ["dep:rig-agent"]
