+++ src/lib.rs
@@ -153,6 +153,9 @@ pub mod prelude {
     // `builder().build()` over the bundled reqwest transport.
     #[cfg(feature = "reqwest")]
     pub use rig_reqwest::prelude::*;
+    // `rmcp_tool*` builder extension traits from `rig-rmcp`.
+    #[cfg(all(feature = "agent", feature = "rmcp", not(target_family = "wasm")))]
+    pub use rig_rmcp::prelude::*;
 }
 
 /// Low-level streaming values plus classic streaming traits.
@@ -187,14 +190,14 @@ pub mod tool {
     // Classic contextual tool API (default runtime). `Tool`/`ToolContext` are
     // the classic contextual trait and its mutable context; none of these
     // collide with the portable exports above.
-    // Native-only, matching every other `rmcp` gate: the module does not exist
-    // on wasm. Reaching this re-export there needs `rig-agent` to have compiled
-    // first, which its own `compile_error!` prevents, so the predicate is
-    // belt-and-braces — but the CI error-count assertion only builds
-    // `-p rig-agent`, so nothing else would catch this one drifting.
+    // MCP tool support from the companion `rig-rmcp` crate (native-only: the
+    // crate root raises a `compile_error!` on wasm, which CI asserts is the
+    // only error). Kept at `rig::tool::rmcp` so existing paths resolve.
     #[cfg(all(feature = "agent", feature = "rmcp", not(target_family = "wasm")))]
     #[cfg_attr(docsrs, doc(cfg(feature = "rmcp")))]
-    pub use rig_agent::tool::rmcp;
+    pub mod rmcp {
+        pub use rig_rmcp::*;
+    }
     #[cfg(feature = "agent")]
     #[cfg_attr(docsrs, doc(cfg(feature = "agent")))]
     pub use rig_agent::tool::{
